+++ src/session_impl.hpp
@@ -160,10 +160,6 @@ namespace sdk {
 
         virtual nlohmann::json cancel_twofactor_reset(const nlohmann::json& twofactor_data) = 0;
 
-        virtual nlohmann::json set_pin(
-            const std::string& mnemonic, const std::string& pin, const std::string& device_id)
-            = 0;
-
         virtual nlohmann::json encrypt_with_pin(const nlohmann::json& details) = 0;
 
         virtual bool encache_blinding_data(const std::string& pubkey_hex, const std::string& script_hex,