+++ subprojects/gdk_rust/gdk_electrum/src/lib.rs
@@ -792,31 +792,6 @@ impl ElectrumSession {
         self.get_account(opt.subaccount)?.get_previous_addresses(opt)
     }
 
-    pub fn set_pin(&self, details: &PinSetDetails) -> Result<PinData, Error> {
-        let agent = self.build_request_agent()?;
-        let manager = PinManager::new(
-            agent,
-            self.network.pin_server_url(),
-            &self.network.pin_manager_public_key()?,
-        )?;
-        let client_key = SecretKey::new(&mut thread_rng());
-        let server_key = manager.set_pin(details.pin.as_bytes(), &client_key)?;
-        let iv = thread_rng().gen::<[u8; 16]>();
-        let cipher = Aes256Cbc::new_from_slices(&server_key[..], &iv).unwrap();
-        let credentials = Credentials {
-            mnemonic: details.mnemonic.clone(),
-        };
-        let plaintext = serde_json::to_vec(&credentials)?;
-        let encrypted = cipher.encrypt_vec(&plaintext);
-
-        let result = PinData {
-            salt: iv.to_hex(),
-            encrypted_data: encrypted.to_hex(),
-            pin_identifier: client_key.to_hex(),
-        };
-        Ok(result)
-    }
-
     pub fn encrypt_with_pin(&self, details: &EncryptWithPinDetails) -> Result<PinData, Error> {
         let agent = self.build_request_agent()?;
         let manager = PinManager::new(