+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 ### Removed
 
 - Removed GA_get_mnemonic_passphrase, callers should use GA_get_credentials.
+- Removed GA_set_pin, callers should use GA_encrypt_with_pin.
 
 
 ## Release 0.0.54 - 2022-05-13